fix(#378): merge dual app delegates to fix startup crash
WikiFSApp declared two @NSApplicationDelegateAdaptor properties with
different types (AppDelegate and QuitConfirmationDelegate, added in #377).
Only one can occupy NSApplication.shared.delegate; accessing the other
triggers an unconditional 'as!' cast that aborts at launch.

The crash occurred ~3s after startup in startStatusItem() when
appDelegate was first read — swift_dynamicCastClassUnconditionalImpl
failed because the actual delegate was QuitConfirmationDelegate.

Fix: fold the quit-confirmation logic (flushPendingSaves,
activeOperationDescription, confirmQuitKey, confirmBeforeQuitting,
applicationShouldTerminate dialog) into the single AppDelegate.
Delete the standalone QuitConfirmationDelegate.swift. Update
GeneralSettingsView to reference AppDelegate.confirmQuitKey.

Kept the existing accessory-mode behavior for
applicationShouldTerminateAfterLastWindowClosed (returns false, drops
to menu-bar-only) over #377's conflicting 'return true' — the quit
dialog still covers all explicit quit paths (Q, Apple menu, Dock,
system shutdown).
